+++ src/docquery/__init__.py
@@ -1,0 +1,3 @@
+# pipeline wraps transformers pipeline with extensions in DocQuery
+# we're simply re-exporting it here.
+from .transformers_patch import pipeline  # noqa
