+++ tests/test_end_to_end.py
@@ -5,9 +5,9 @@ import pytest
 from pydantic import BaseModel
 from transformers.testing_utils import nested_simplify
 
+from docquery import pipeline
 from docquery.document import load_document
 from docquery.ocr_reader import TesseractReader
-from docquery.pipeline import get_pipeline
 
 
 CHECKPOINTS = {
@@ -79,18 +79,18 @@ EXAMPLES = [
 @pytest.mark.parametrize("model", CHECKPOINTS.keys())
 def test_impira_dataset(example, model):
     document = load_document(example.path)
-    pipeline = get_pipeline(checkpoint=CHECKPOINTS[model])
+    pipe = pipeline("document-question-answering", checkpoint=CHECKPOINTS[model])
     for qa in example.qa_pairs:
-        resp = pipeline(question=qa.question, **document.context, top_k=1)
+        resp = pipe(question=qa.question, **document.context, top_k=1)
         assert nested_simplify(resp, decimals=4) == qa.answers[model]
 
 
 def test_run_with_choosen_OCR_str():
     example = EXAMPLES[0]
     document = load_document(example.path, "tesseract")
-    pipeline = get_pipeline(checkpoint=CHECKPOINTS["LayoutLMv1"])
+    pipe = pipeline("document-question-answering", checkpoint=CHECKPOINTS["LayoutLMv1"])
     for qa in example.qa_pairs:
-        resp = pipeline(question=qa.question, **document.context, top_k=1)
+        resp = pipe(question=qa.question, **document.context, top_k=1)
         assert nested_simplify(resp, decimals=4) == qa.answers["LayoutLMv1"]
 
 
@@ -98,7 +98,7 @@ def test_run_with_choosen_OCR_instance():
     example = EXAMPLES[0]
     reader = TesseractReader()
     document = load_document(example.path, reader)
-    pipeline = get_pipeline(checkpoint=CHECKPOINTS["LayoutLMv1"])
+    pipe = pipeline(checkpoint=CHECKPOINTS["LayoutLMv1"])
     for qa in example.qa_pairs:
-        resp = pipeline(question=qa.question, **document.context, top_k=1)
+        resp = pipe(question=qa.question, **document.context, top_k=1)
         assert nested_simplify(resp, decimals=4) == qa.answers["LayoutLMv1"]
